+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.17.17
 
 * added automatic enum parsing to satella's JSONEncoder
+* Satella's JSON file handling will internally use Satella's JSONEncoder